% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allom.predict.R
\name{allom.predict}
\alias{allom.predict}
\title{allom.predict}
\usage{
allom.predict(
  object,
  dbh,
  pft = NULL,
  component = NULL,
  n = NULL,
  use = "Bg",
  interval = "prediction",
  single.tree = FALSE
)
}
\arguments{
\item{object}{Allometry model object. Option includes
\describe{
  \item{'list of mcmc'}{ - mcmc outputs in a list by PFT then component}
  \item{'vector of file paths'}{ - path(s) to AllomAve RData files}
  \item{'directory where files are located}{ - }
}}

\item{dbh}{Diameter at Breast Height (cm)}

\item{pft}{Plant Functional Type. Needs to match the name used in AllomAve.
Can be NULL if only one PFT/species exists, otherwise needs to the same length as dbh}

\item{component}{Which component to predict. Can be NULL if only one component was analysed in AllomAve.}

\item{n}{Number of Monte Carlo samples. Defaults to the same number as in the MCMC object}

\item{use}{c('Bg','mu','best')}

\item{interval}{c('none','confidence','prediction') default is prediction}

\item{single.tree}{logical: Is this a DBH time series from one individual tree?
If TRUE, will use a fixed error for all draws.}
}
\value{
matrix of Monte Carlo predictions that has n rows and one column per DBH
}
\description{
Function for making tree-level Monte Carlo predictions
from allometric equations estimated from the PEcAn allometry module
}
\examples{
\dontrun{
object <- "~/Dropbox//HF C Synthesis/Allometry Papers & Analysis/"
dbh <- seq(10, 50, by = 5)
mass <- allom.predict(object, dbh, n = 100)
}

}
\author{
Michael Dietze, Christy Rollinson
}
