\name{AllomUnitCoef}
\alias{AllomUnitCoef}
\title{AllomUnitCoef}
\usage{
  AllomUnitCoef(x, tp = NULL)
}
\arguments{
  \item{x}{units: mm, cm, cm2, m, in, g, kg, lb, Mg}

  \item{tp}{diameter type, leave NULL if DBH. Options:
  'd.b.h.^2','cbh','crc'}
}
\description{
  converts length units FROM cm TO specified units converts
  mass units TO kg FROM specificed units
}

