% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.allometry.object.R
\name{as.allometry.object}
\alias{as.allometry.object}
\title{Make an 'allometry' object.}
\usage{
as.allometry.object(intercept, slope, x_mean, y_mean)
}
\arguments{
\item{vector}{of allometric intercepts}

\item{vector}{of allometric slopes}

\item{vector}{of mean body size}

\item{vector}{of mean trait size}
}
\value{
a 'allometry' object.
}
\description{
This function combines data into an 'allometry' object.
}
\references{
Voje, K. L., Hansen, T. F., Egset, C. K., Bolstad, G. H., & Pélabon, C. (2014). Allometric constraints and the evolution of allometry. \emph{Evoluton} 68:866–885.
}
\author{
Kjetil L. Voje
}
