% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint.slope.R
\name{constraint.slope}
\alias{constraint.slope}
\title{Constraint on trait variance when keeping slope fixed}
\usage{
constraint.slope(allometry.object)
}
\arguments{
\item{allo}{an allometry object}
}
\value{
Constraint on trait variance when keeping slope fixed
}
\description{
Constraint on trait variance when keeping slope fixed
}
\details{
The constraint is quantified by calculating the decrease in trait variance that results when conditioning on one or more constraining parameters.
}
\references{
Voje, K. L., Hansen, T. F., Egset, C. K., Bolstad, G. H., & Pélabon, C. (2014). Allometric constraints and the evolution of allometry. \emph{Evoluton} 68:866–885.
}
\author{
Kjetil L. Voje
}
