% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_b_x.a.R
\name{cov_b_x.a}
\alias{cov_b_x.a}
\title{Calculating the covariance between slope and size conditioned on intercept}
\usage{
cov_b_x.a(allometry.object)
}
\arguments{
\item{allo}{an allometry object}
}
\value{
The the covariance between slope and size conditioned on intercept
}
\description{
Calculating the covariance between slope and size conditioned on intercept
}
\references{
Voje, K. L., Hansen, T. F., Egset, C. K., Bolstad, G. H., & Pélabon, C. (2014). Allometric constraints and the evolution of allometry. \emph{Evoluton} 68:866–885.
}
\author{
Kjetil L. Voje
}
