% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earwigs}
\alias{earwigs}
\title{Morphological measurement taken from 113 maritime earwigs (Anisolabis maritima); 68 males and 45 females.}
\format{
A data frame with twelve variables:
\describe{
\item{\code{id}}{ID (integer; starting from 1, ending at 115; 11 and 12 are missing)}
\item{\code{sex}}{Sex (factor; male or female)}
\item{\code{head_width}}{Head width (mm)}
\item{\code{head_length}}{Head length (mm)}
\item{\code{thorax_width}}{Thorax width (mm)}
\item{\code{thorax_length}}{Thorax length (mm)}
\item{\code{abdomen_width}}{Abdomen width (mm)}
\item{\code{abdomen_length}}{Abdomen length (mm)}
\item{\code{right_forcep_length}}{Right forcep length (mm)}
\item{\code{left_forcep_length}}{Left forcep length (mm)}
\item{\code{right_forcep_width}}{Right forcep width (mm)}
\item{\code{left_forcep_width}}{Left forcep width (mm)}
}

For further details, see Tomoki Matsuzawa, Junji Konuma, Positive allometry in the forceps of the female earwig Anisolabis maritima (Dermaptera: Anisolabididae), Biological Journal of the Linnean Society, Volume 145, Issue 2, June 2025, blaf031, \doi{10.1093/biolinnean/blaf031}.
}
\usage{
earwigs
}
\description{
Morphological measurement taken from 113 maritime earwigs (Anisolabis maritima); 68 males and 45 females.
}
\keyword{datasets}
