% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.all.constraints.R
\name{fit.all.constraints}
\alias{fit.all.constraints}
\title{Calculating all possible (six) constraints at the same time.}
\usage{
fit.all.constraints(allometry.object)
}
\arguments{
\item{allo}{an allometry object}
}
\value{
First part of the output gives the observed trait variance and the number of allometries in the dataset. The second
part of the output gives predicted total trait variance that remains when conditioning on different parameters in percent. 
E.g. a value of 50 means 50% of the predicted trait variance remains after conditioning on specific parameter(s).
}
\description{
Calculating all possible (six) constraints at the same time.
}
\details{
A wrapper function for investigating all possible (six) combinations of constraints at the same time.
}
\references{
Voje, K. L., Hansen, T. F., Egset, C. K., Bolstad, G. H., & Pélabon, C. (2014). Allometric constraints and the evolution of allometry. \emph{Evoluton} 68:866–885.
}
\author{
Kjetil L. Voje
}
