% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allom.predict.R
\name{load.allom}
\alias{load.allom}
\title{load.allom}
\usage{
load.allom(object)
}
\arguments{
\item{object}{Allometry model object. Option includes
\describe{
  \item{'vector of file paths'}{ - path(s) to AllomAve RData files}
  \item{'directory where files are located}{ - }
}}
}
\value{
mcmc outputs in a list by PFT then component
}
\description{
loads allom files
}
\examples{
\dontrun{
object <- "~/Dropbox//HF C Synthesis/Allometry Papers & Analysis/"
allom.mcmc <- load.allom(object)
}

}
\author{
Michael Dietze
}
