% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_a.x.R
\name{var_a.x}
\alias{var_a.x}
\title{Calculating the variance in intercepts conditioned on size}
\usage{
var_a.x(allometry.object)
}
\arguments{
\item{allo}{an allometry object}
}
\value{
The variance in intercepts conditioned on size
}
\description{
Calculating the variance in intercepts conditioned on size
}
\references{
Voje, K. L., Hansen, T. F., Egset, C. K., Bolstad, G. H., & Pélabon, C. (2014). Allometric constraints and the evolution of allometry. \emph{Evoluton} 68:866–885.
}
\author{
Kjetil L. Voje
}
