% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_b.a.R
\name{var_b.a}
\alias{var_b.a}
\title{Calculating the variance in slope conditioned on intercept}
\usage{
var_b.a(allometry.object)
}
\arguments{
\item{allo}{an allometry object}
}
\value{
The variance in slope conditioned on intercept
}
\description{
Calculating the variance in slope conditioned on intercept
}
\references{
Voje, K. L., Hansen, T. F., Egset, C. K., Bolstad, G. H., & Pélabon, C. (2014). Allometric constraints and the evolution of allometry. \emph{Evoluton} 68:866–885.
}
\author{
Kjetil L. Voje
}
