% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_x.b.R
\name{var_x.b}
\alias{var_x.b}
\title{Calculating the variance in size conditioned on slope}
\usage{
var_x.b(allometry.object)
}
\arguments{
\item{allo}{an allometry object}
}
\value{
The variance in size conditioned on slope
}
\description{
Calculating the variance in size conditioned on slope
}
\references{
Voje, K. L., Hansen, T. F., Egset, C. K., Bolstad, G. H., & Pélabon, C. (2014). Allometric constraints and the evolution of allometry. \emph{Evoluton} 68:866–885.
}
\author{
Kjetil L. Voje
}
